✨ FIX user sign up
- fix h2 database and id

+++ Backend/3laka/src/main/java/com/ssafy/laka/domain/Comment.java
@@ -17,7 +17,7 @@ public class Comment {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    public Long commentId;
+    public int commentId;
 
     @ManyToOne(targetEntity = User.class, fetch = FetchType.LAZY)
     @JoinColumn(name = "user_id")
