+++ Backend/3laka/src/main/java/com/ssafy/laka/domain/Friend.java
@@ -18,7 +18,7 @@ public class Friend {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long friendId;
+    private int friendId;
 
     @ManyToOne(targetEntity = User.class, fetch = FetchType.LAZY)
     @JoinColumn(name="sender")
